Separating Skills from CharacterBase / Version up ReduceSkillCooldown method
